User memory access: handle invalid memory addresses with EFAULT
This change implements a functionality similar to
setjmp()/longjmp(), i.e. the ability for kernel code to jump to an
arbitrary location within the current function call stack. This
functionality is used to handle page fault exceptions that may be
addresses, so that an appropriate error code (usually -EFAULT) is
returned to userspace if a page fault cannot be resolved.
The implementation relies on a simplified context frame that is
being added to the kernel_context struct; this additional frame
(referred to as the error frame) contains all callee-saved
registers as listed in the function call conventions for a given
architecture, plus an instruction pointer.
The error frame for the current context is filled by calling
`err_frame_save` (which returns a false value when called
directly), and is applied to a given context by calling
err_frame_apply(), which makes the context "jump" to the return
address of the latest err_frame_save() invocation, which returns
a true value when jumped to in this way. When a piece of code is
done accessing user memory, it clears the error frame of the
current context by calling context_clear_err(). err_frame_apply()
is called by both the Unix fault handler and the pending fault
completion if a page fault cannot be resolved.
The copy_to_user() and copy_from_user() functions (which copy data
between user and kernel memory) use the above functionality to
return a boolean value that indicates whether the copy succeeded,
i.e. whether any page faults generated when accessing user memory
could be resolved with an appropriate virtual address mapping.
Similarly, get_user_value() and set_user_value() are utility macros
that copy simple variables between user and kernel memory.
The sg_fault_in() function has been amended to return a boolean
value, which is checked in pagecache_write_sg_finish() so that an
approriate error is returned if a page fault cannot be resolved;
this sg_fault_in() function call is now made before acquiring the
pagecache node lock, in order to avoid suspending the current
context with spinlocks held.
The Unix code that accesses user memory has been amended to
properly detect unresolvable page faults and return an error code.
The err_frame_save() function is annotated with
__attribute__((returns_twice)) in order to prevent compiler
optimizations that could make the value of local variables
indeterminate after an error frame is applied to the running
context.
The fault_in_user_memory() function has been simplified to take a
boolean argument indicating whether a given memory range should be
writable, and traverses the process virtual memory map only if this
argument is true (if a memory range needs read-only access, there
is no need to traverse the memory map to check the protection
flags: just touching each page in the range is enough).
The resolve_dir() macro has been ameded to fault in the provided
file path string and return -EFAULT if the string could not be
faulted in. In some syscall handlers (some of which call
`resolve_dir`), redundant calls to fault_in_user_string() have been
removed.
Runtime tests exercising different syscalls have been enhanced to
test proper handling of invalid memory addresses.

+++ test/runtime/unlink.c
@@ -8,6 +8,8 @@
 #include <string.h>
 #include <sys/vfs.h>
 
+#define FAULT_ADDR  ((void *)0xBADF0000)
+
 static void test_unlink(const char *path)
 {
     struct stat s;
@@ -31,6 +33,11 @@ static void test_unlink(const char *path)
 static void test_unlinkat(int dirfd)
 {
     struct stat s;
+
+    if ((unlinkat(dirfd, FAULT_ADDR, 0) != -1) || (errno != EFAULT)) {
+        printf("unlinkat test with faulting path failed\n");
+        exit(EXIT_FAILURE);
+    }
 
     int fd = openat(dirfd, "file", O_CREAT, S_IRWXU);
     if (fd < 0) {
@@ -101,6 +108,10 @@ int main(int argc, char **argv)
     int fd;
     struct stat s;
 
+    if ((unlink(FAULT_ADDR) != -1) || (errno != EFAULT)) {
+        printf("unlink test with faulting path failed\n");
+        exit(EXIT_FAILURE);
+    }
     if ((unlink("") != -1) || (errno != ENOENT)) {
         printf("empty path unlink test failed\n");
         exit(EXIT_FAILURE);
